+++ src/js/modules/ParagraphModule.js
@@ -9,57 +9,79 @@ const ParagraphModule = (paragraph) => {
     /* Global Variables */
 
     /* array of all span elements containing words/special chars in the paragraph */
-    let spansInParagraph; 
-    
+    let spansInParagraph;
+
     /* 
      - array of value (in px) of current word distance from top of page
        to previous word distance from top of page 
      - added 0 to first element to length would match spansInParagraph.lenght (see calculate lines)
     */
-    let topOfWordDifferences = [0];
-    
-    /* array of Line modules */
-    let linesInParagraph = []; 
+    let topOfSpanDifferences = [0];
+
+    /* array of Line modules (eventually), currently its a 2d array of span elems */ 
+    let linesInParagraph = [];
 
     /* Module fucntions */
 
     const splitIntoSpans = () => {
 
-        Splitting( { target: paragraph, by: "paragraphSplitPlugin" } )
+        Splitting({ target: paragraph, by: "paragraphSplitPlugin" })
         // copy all "spannified" words into array
         spansInParagraph = $(paragraph).find("span.word, span.whitespace");
 
     }
 
-    // break down into different functions
+    /*
+     - Calls the necessary functions to break a paragraph into LineModules
+    */
     const calculateLines = () => {
 
-        // added 0 to first element of topOfWordDifferences so it would have the same length as spansInParagraph - this loop MUST start at i = 1
-        // calculates the difference between the offset top of the current word and the offset top of the previous word
-        
-        for (let i = 1; i < spansInParagraph.length; i++) {
-         
-            let currentWordOffsetTop = $(spansInParagraph[i]).offset().top; 
-            let previousWordOffsetTop = $(spansInParagraph[i - 1]).offset().top; 
-            
+        populateDifferencesArray();
 
-            topOfWordDifferences.push(
-                Math.round(currentWordOffsetTop - previousWordOffsetTop)
+        populateLineModuleArray();
+
+    }
+
+    /*
+     - Populates an array (topOfSpanDifferences) with numbers that represent the difference between the offset top of the current span element and the offset top of the previous span element
+     - Offset top is the distance from the top of the page to the top of the span element 
+     - topOfSpanDifferences starts at idx 1 because the loop needs to start at 1, and the arr needs to be the same length as spansInParagraph
+    */
+    const populateDifferencesArray = () => {
+
+        for (let i = 1; i < spansInParagraph.length; i++) {
+
+            let currentSpanOffsetTop = $(spansInParagraph[i]).offset().top;
+            let previousSpanOffsetTop = $(spansInParagraph[i - 1]).offset().top;
+
+
+            topOfSpanDifferences.push(
+                Math.round(currentSpanOffsetTop - previousSpanOffsetTop)
             );
         }
+
+    }
+
+    /*
+     - Uses the differences array to check for line breaks
+     - There is a line break if the difference is larger than the line height
+     - Using startOfLineIdx to track the first word in every line, all elements between that index and the index where the line breaks is a line
+     - Creates a LineModule for each line, and adds it to linesInParagraph
+    */
+    const populateLineModuleArray = () => {
 
         let startOfLineIdx = 0; // used to keep track of the first word in a line
         let lineHeightOfPreviousWord = 0;
 
-        // use differences calculated in previous loop to find all of the lines in the paragraph
-        for (let i = 0; i < topOfWordDifferences.length; i++) {
+        // use differences array to find all of the lines in the paragraph
+        for (let i = 0; i < topOfSpanDifferences.length; i++) {
 
             if (i > 0) {
                 lineHeightOfPreviousWord = $(spansInParagraph[i - 1]).outerHeight();
             }
 
             // if there is a difference that is bigger than the lineHeight, that means there is a break in the lines - pushes all elements inbetween startOfLineIdx and the linebreak (the line) 
-            if (topOfWordDifferences[i] > lineHeightOfPreviousWord) {
+            if (topOfSpanDifferences[i] > lineHeightOfPreviousWord) {
 
                 linesInParagraph.push(
                     spansInParagraph.slice(startOfLineIdx, i)
@@ -71,8 +93,9 @@ const ParagraphModule = (paragraph) => {
 
         }
 
+        // account for the last line in the paragraph
         linesInParagraph.push(
-            spansInParagraph.slice(startOfLineIdx, topOfWordDifferences.length)
+            spansInParagraph.slice(startOfLineIdx, topOfSpanDifferences.length)
         );
 
     }